refact: inline SemanticsNode.isDisplayed
